refactor: define project and package name
This commit changes the project name to "wizard", sets the package name
to "com.github.wizard", sets the package name for the protobuf generated
files to "com.github.wizard.api". It also changes the names of the
subprojects from "app" to "wizard-app" and from "wizzardserver" to
"wizard-server".
