+++ .idea/.name
@@ -1,1 +1,1 @@
-My Application
+wizard
