+++ .idea/gradle.xml
@@ -11,7 +11,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/wizzardserver" />
+            <option value="$PROJECT_DIR$/server" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
