+++ app/build.gradle
@@ -7,7 +7,6 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.example.myapplication"
         minSdk 21
         targetSdk 32
         versionCode 1
