+++ app/src/main/java/com/github/wizard/GamePlayActivity.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.github.wizard;
 
 import android.content.Intent;
 import android.os.Bundle;
