+++ protobuf/api.proto
@@ -1,10 +1,8 @@
 syntax = "proto3";
 option java_multiple_files = true;
-option java_package = "server";
-option java_outer_classname = "ServerProto";
-option objc_class_prefix = "HLW";
+option java_package = "com.github.wizard.api";
 
-package server;
+package api;
 
 service GameStarter{
   rpc startGame(StartRequest)returns(StartReply) {}
