+++ server/src/main/java/com/github/wizard/Game.java
@@ -1,4 +1,4 @@
-package com.example.wizzardserver;
+package com.github.wizard;
 
 public class Game {
     public final int gameId;
